Harden sequence tracking and publisher send state after review
Cross-cutting review of the Phase 3.5 work found interaction bugs
between the new mechanisms; this closes them.

The per-writer DataSetMessage sequence counter now rolls over at the
wire DataType maximum per 7.2.3 - UInt16 for UADP, UInt32 for JSON
(Table 185). The previous mapping-agnostic 16-bit counter wedged
Milo-to-Milo JSON streams permanently stale at the 65535 wrap against
the subscriber's UInt32 window. The UADP encoder now rejects
externally built drafts whose transmitted sequence number exceeds the
UInt16 wire range instead of silently truncating.

The per-reader NetworkMessage window is observed once per matched
NetworkMessage even when no carried DataSetMessage passes the
reader's writer filter; previously a quiet writer beside busy
siblings starved a filtered reader's window until resumed traffic
classified invalid. Publisher restarts now recover: a keep-alive
whose carried next-expected value falls outside the window reseeds it
per 7.2.4.5.8 (the publisher is authoritative about its own counter),
and when no receive timeout is armed - where the 7.2.3 record-discard
rule is undefined - sixteen consecutive rejections discard the stream
record so the next message reseeds. Reconfigure restarting a writer's
sequence at zero is documented together with those recovery paths.

Publisher send state is now transmit-time: lastSentNanos commits at
channel hand-off rather than draft creation, so a writer whose data
messages are skipped (for example over maxNetworkMessageSize) lapses
into keep-alives instead of going silent, and writers in Error emit
no keep-alives at all (6.2.6.3 - a keep-alive asserts the writer is
still active, and Error-state keep-alives masked activation
failures). Delta baselines were already invalidated on skipped sends;
tests now pin that. PubSubInteropTool prints the wire sequence
numbers and surfaces partial-decode failures in raw mode.

+++ opc-ua-sdk/sdk-pubsub/src/main/java/org/eclipse/milo/opcua/sdk/pubsub/json/JsonNetworkMessageEncoder.java
@@ -392,8 +392,9 @@ final class JsonNetworkMessageEncoder {
       writer.name("WriterGroupName").value(context.writerGroup().getName());
     }
     if (messageMask.getSequenceNumber()) {
-      // emitted as UInt32 (Table 185); the engine counter is UInt16 and wraps at 65535
-      writer.name("SequenceNumber").value(draft.sequenceNumber().intValue());
+      // emitted as UInt32 (Table 185); the engine's JSON sequence counter spans the full UInt32
+      // range and rolls over at 2^32 (§7.2.3 rollover at the DataType maximum)
+      writer.name("SequenceNumber").value(draft.sequenceNumber().longValue());
     }
     if (messageMask.getMetaDataVersion()) {
       ConfigurationVersionDataType version = draft.configurationVersion();
